+++ src/js/page/terminal/commands/__all__.mjs
@@ -15,6 +15,7 @@ import cmdLoad from './load';
 import cmdPrint from './print';
 import cmdQuit from './quit';
 import cmdToggleTerm from './toggle_term';
+import cmdInput from './input';
 import type VMachine from '@trash/vmachine';
 
 export default function (vm: VMachine) {
@@ -32,4 +33,5 @@ export default function (vm: VMachine) {
   vm.registerCommand('toggle_term', cmdToggleTerm());
   vm.registerCommand('dis', cmdDis());
   vm.registerCommand('genfile', cmdGenFile());
+  vm.registerCommand('input', cmdInput());
 }